pty: Fix shutdown issues
Don't terminate the task/pty on Windows on shutdown, instead just close
the files.  Closing a windows pty will block until the process
terminates.

Add a timeout waiting for pty shutdown.

+++ lib/gensio_win.c
@@ -2631,10 +2631,6 @@ win_pty_shutdown(struct gensio_iod_win_pty *piod, bool force)
 	    return rv;
 	piod->read = NULL;
     }
-    if (piod->ptyh) {
-	ClosePseudoConsole(piod->ptyh);
-	piod->ptyh = NULL;
-    }
     return 0;
 }
 
@@ -2648,7 +2644,8 @@ win_pty_close(struct gensio_iod_win *wiod, bool force)
     rv = win_pty_shutdown(piod, force);
     if (rv)
 	goto out;
-    TerminateProcess(piod->child, 0);
+    if (piod->child)
+	TerminateProcess(piod->child, 0);
     if (piod->watch_thread) {
 	LeaveCriticalSection(&wiod->lock);
 	WaitForSingleObject(piod->watch_thread, INFINITE);
